Rename tempFooter to Footer
