+++ src/app/layout.tsx
@@ -5,7 +5,7 @@ import ReactQueryProvider from './_lib/providers/ReactQueryProvider';
 import StoreProvider from './_lib/providers/StoreProvider';
 import NextUiProvider from './_lib/providers/NextUiProvider';
 import Navigation from './_components/navigation/Navigation';
-import Footer from './_components/footer/tempFooter';
+import Footer from './_components/footer/Footer';
 import './globals.css';
 
 const roboto = Roboto({
